update shared test setup
