+++ .vscode/launch.json
@@ -62,7 +62,9 @@
             "module": "pytest",
             "args": [
                 "-v",
-                "tests/test_router_quote.py"
+                "tests/test_router_quote.py",
+                "-n",
+                "4"
             ],
             "console": "integratedTerminal",
             "justMyCode": true,
