+++ tests/constants.py
@@ -1,5 +1,3 @@
-import setup
-
 # Misc constants
 UOSMO = "uosmo"
 USDC = 'ibc/498A0751C798A0D9A389AA3691123DADA57DAA4FE165D5C75894505B876BA6E4'
